Transfer UI related fields and methods to GUIManager
